% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krkFun.R
\name{bracken_read_report}
\alias{bracken_read_report}
\title{Reads bracken reports}
\usage{
bracken_read_report(report.file)
}
\arguments{
\item{report.file}{name of file with bracken report output.}
}
\value{
A data.frame with the \code{bracken} results for each taxon. It has the columns:
\itemize{
  \item name. The name of this taxon in the taxonomy.
  \item tax_id. The taxonomy identifier of this taxon in the taxonomy.
  \item rank. The rank of this taxon in the taxonomy.
  \item tax_count_krk. The number of reads classified to this taxon by \code{kraken2}.
  \item added_brk. The additional reads assigned by \code{bracken}.
  \item tax_count. The final number of reads assigned to this taxon.
  \item fraction. The fraction of the total number of reads assigned to this taxon.
}
}
\description{
Reads the report output from \code{bracken}.
}
\details{
The report output from bracken is a table with one row for each
taxon with assigned reads. This is the file produced by \code{bracken} using
the \code{-o} option.

Note that by the \code{-w} option \code{bracken} will output a report in
\code{kraken2} format, see \code{\link{kraken2_read_report}}.
}
\examples{

}
\seealso{
\code{\link{kraken2_read_report}}.
}
\author{
Lars Snipen.
}
