% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{branch_list2qiime}
\alias{branch_list2qiime}
\title{Branch list to QIIME format}
\usage{
branch_list2qiime(branch.lst)
}
\arguments{
\item{branch.lst}{a list of branches.}
}
\value{
The vector of texts, one for each element in \code{branch.lst}.
}
\description{
Creates QIIME formatted texts from a branch list.
}
\details{
A QIIME formatted branch is a text with the format 

\code{"d__domain;p__phylum;c__class;o__order;f__family;g__genus;s__species"}

where the words domain, phylum, etc are replaced by taxon names. This function converts
a list of N branches to N such texts.

NOTE 1: This is only meaningful if the \code{branch.lst} contains names instead of tax_id's, see
\code{\link{branch_taxid2name}}.

NOTE 2: The ranks are fixed to the ones listed above, where domain is the same as superkingdom in
the NCBI taxonomy.
}
\examples{

}
\seealso{

}
\author{
Lars Snipen.
}
