% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{branch_list2table}
\alias{branch_list2table}
\alias{branch_table2list}
\title{Branch list na dtable conversion}
\usage{
branch_list2table(
  branch.lst,
  ranks = c("superkingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{branch.lst}{a list of branches.}

\item{ranks}{texts specifying the ranks to keep and their ordering.}

\item{branch.tbl}{a table of branches.}
}
\value{
The \code{branch_list2table} returns a tibble with one row for each branch and a column for
each rank. The cells contain the tax_id's or names in the \code{branch.lst}.

The \code{branch_table2list} returns a list where each element contains a named vector of tax_id or 
texts, and the name of each element is its rank in the taxonomy.
}
\description{
Turns a list of branches into a table with, or vice versa.
}
\details{
Instead of having branches in a list, it is convenient in R to have data in tables. The
\code{branch_list2table} converts a list of branches into a table, with one row for each branch
and a column for each rank. Since branches may contain different ranks, they must first be pruned
to have the exact same ranks in the exact same order. This is done by \code{\link{branch_prune}}
inside this function. The ranks specified in \code{ranks} are the columns in the table returned here.

The function \code{branch_table2list} converts back again, i.e. from a table to a list.
}
\examples{

}
\seealso{

}
\author{
Lars Snipen.
}
