% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{branch_prune}
\alias{branch_prune}
\title{Filtering ranks in branches}
\usage{
branch_prune(
  branch.lst,
  ranks = c("superkingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{branch.lst}{a list of branches.}

\item{ranks}{texts specifying the ranks to keep and their ordering.}
}
\value{
A new branch-list with pruned branches all containing the exact same ranks.
}
\description{
Keeps only specified ranks in specified ordering in all branches.
}
\details{
Branches in the taxonomy tree may have many different ranks or levels. This function is used
to prune them all to have the same set of ranks. Ranks may be missing in some branches.This function
will then fill in NA in these cells, ensuring all branches have the exact same ranks in the exact same
ordering. This is convenient for turning the list into a table, using \code{\link{as_tibble}}.
}
\author{
Lars Snipen.
}
