% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{branch_retrieve}
\alias{branch_retrieve}
\title{Retrieves branches}
\usage{
branch_retrieve(leaf_tax_id, nodes.dmp)
}
\arguments{
\item{leaf.tax.id}{a vector of tax_id (integers) of the branch leaf/leaves.}

\item{nodes.dmp}{a nodes.dmp table.}
}
\value{
A branch list, which is simply a list containing named vectors of tax_id integers, from the root to
the leaf of each branch. The name of each element is its rank. You may replace the tax_id integers by the 
taxon names using \code{\link{branch_taxid2name}}.
}
\description{
Retrieves branches in the taxonomy tree.
}
\details{
This function retrieves the branch from the taxonomy tree that ends at the \code{leaf.tax.id}
and starts at the root of tree defined in the table \code{nodes.dmp} (see \code{\link{read_nodes_dmp}}).

NB! Only the root node in nodes.dmp must have its own tax_id as parent_tax_id! This is the criterion for
ending a branch.

Multiple leaves may be given as argument, resulting in multiple branches retrieved.
}
\seealso{
\code{\link{branch_taxid2name}}, \code{\link{subset_clade}}.
}
\author{
Lars Snipen.
}
