% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{branch_taxid2name}
\alias{branch_taxid2name}
\alias{branch_taxid2name,}
\alias{branch_name2taxid}
\title{Replace tax_id with name}
\usage{
branch_taxid2name(branch.lst, names.dmp)
}
\arguments{
\item{branch.lst}{a list of branches.}

\item{names.dmp}{a names.dmp table.}
}
\value{
A list containing named vectors of either texts (\code{branch_taxid2name}) or integers 
(\code{branch_name2taxid}), from the root to the leaf of each branch.
The name of each element is its rank.
}
\description{
Converts vectors of tax_id to vectors of names, or vice versa.
}
\details{
This function is used to convert the tax_id's to name_txt's in a branch-list, or vice versa. See 
\code{\link{branch_retrieve}} for more about branch-lists.

\code{branch_taxid2name}: The \code{names.dmp$name_txt} may contain many names for each tax_id, and the
argument \code{name.class} is used to select the type of name to use from the column \code{names.dmp$name_class}.

\code{branch_name2taxid}: The matching of names is exact (using \code{\link{match}}), which means only the first
occurrence of the name_txt is used. You are responsible for using a \code{names.dmp} with unique texts in
\code{names.dmp$name_txt}.

See \code{\link{read_names_dmp}} for more about \code{names.dmp} tables.
}
\author{
Lars Snipen.
}
