% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{getDomain}
\alias{getDomain}
\alias{getPhylum}
\alias{getClass}
\alias{getOrder}
\alias{getFamily}
\alias{getGenus}
\alias{getTag}
\alias{getTaxonomy}
\title{Extractor functions for QIIME taxonomy}
\usage{
getDomain(header)
getPhylum(header)
getClass(header)
getOrder(header)
getFamily(header)
getGenus(header)
getTag(header)
getTaxonomy(header)
}
\arguments{
\item{header}{A vector of texts, the \code{Header} column from reading a FASTA file,
containing taxonomy information in the proper format.}
}
\value{
A vector containing the sub-texts extracted from each \code{header} text, but 
\code{getTaxonomy} returns a table with the full taxonomy, one row for each input \code{header}
}
\description{
Extracting taxonomic information from FASTA headers.
}
\details{
The ConTax data sets (see package \code{microcontx}) are FASTA files 
where the \code{Header} contains texts according to a strict format inherited from QIIME:

It always starts with a short text, a Tag, which is a unique identifier for every sequence.
The function \code{getTag} will extract this from the \code{header}.

After the Tag follows one or more tokens. One of these tokens is a string with the
following format, inherited from QIIME:

"k__<...>;p__<...>;c__<...>;o__<...>;f__<...>;g__<...>;"

where <...> is some proper text. Here is an example of a proper string:

"k__Bacteria;p__Firmicutes;c__Bacilli;o__Bacillales;f__Staphylococcaceae;g__Staphylococcus;"

The functions \code{getDomain}, ..., \code{getGenus} extracts the
corresponding information from the \code{header}. \code{getTaxonomy}
combines all taxonomy extractors, combines these in a table
and imputes missing taxa with parent taxa.
}
\examples{
data(contax.trim)
getTag(contax.trim$Header)
getGenus(contax.trim$Header)
getPhylum(contax.trim$Header)

}
\seealso{
\code{\link{contax.trim}}, \code{\link{medoids}}.
}
\author{
Lars Snipen.
}
