% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krkFun.R
\name{kraken2_read_report}
\alias{kraken2_read_report}
\title{Reads \code{kraken2} reports}
\usage{
kraken2_read_report(report.file)
}
\arguments{
\item{report.file}{name of file with \code{kraken2} report output.}
}
\value{
A data.frame with the \code{kraken2} results for each taxon. It has the columns:
\itemize{
  \item percent. The percentage of the reads classified to this taxon.
  \item clade_count. The total number of reads classified to the clade originating from this taxon.
  \item tax_count. The number of reads assigned directly to this taxon.
  \item rank. The rank of this taxon in the taxonomy.
  \item tax_id. The taxonomy identifier of this taxon in the taxonomy.
  \item name. The name of this taxon in the taxonomy.
}
}
\description{
Reads the report output from \code{kraken2}.
}
\details{
The report output from kraken2 is a table with one row for each
taxon detected by kraken2. Run \code{kraken2} using the \code{--report}
option to get this convenient summary table.

Note that the software \code{bracken} may also output a report in this format.
}
\seealso{
\code{\link{kraken2_read_table}}, \code{\link{bracken_read_report}}.
}
\author{
Lars Snipen.
}
