% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krkFun.R
\name{kraken2_read_table}
\alias{kraken2_read_table}
\title{Reads raw \code{kraken2} output}
\usage{
kraken2_read_table(krk.file, filter = TRUE)
}
\arguments{
\item{krk.file}{name of file with raw \code{kraken2} output.}

\item{filter.}{returns only classified reads if TRUE.}
}
\value{
A data.frame with the kraken2 results for each read. It has the columns:
\itemize{
  \item status. C for classified reads, U for unclassified (if \code{filter = TRUE} only C).
  \item read_id. Short text identifying each read.
  \item name. Name of taxon, if classified.
  \item length. Length of read.
  \item tax_count. The k-mer counts for the read.
}
}
\description{
Reads the raw output table from \code{kraken2}.
}
\details{
The raw output from \code{kraken2} is a huge table with one row for each
read classified.

Note that in most cases only the report, produced by running \code{kraken2} with
the \code{--report} otpion, is what you want. Then use the function 
\code{\link{kraken2_read_report}}.
}
\examples{

}
\seealso{
\code{\link{kraken2_read_report}}.
}
\author{
Lars Snipen.
}
