% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclass-package.R
\docType{package}
\name{microclass-package}
\alias{microclass-package}
\alias{microclass}
\title{Methods for 16S based taxonomic classification of prokaryotes}
\usage{
microclass()
}
\description{
The package provides functions for assigning 16S sequence data to a
taxonomic level in the tree-of-life for prokaryotes.
}
\details{
\tabular{ll}{
Package: \tab microclass\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-08-31\cr
License: \tab GPL-2\cr
}
}
\author{
Hilde Vinje, Kristian Hovde Liland, Lars Snipen.\cr
Maintainer: Lars Snipen <lars.snipen@nmbu.no>
}
\keyword{package}
