% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclr.R
\name{microCLR}
\alias{microCLR}
\title{Centred log-ratio transform}
\usage{
microCLR(readcount.mat, n.pseudo = 0)
}
\arguments{
\item{readcount.mat}{matrix with readcount data.}

\item{n.pseudo}{number of pseudo-readcounts to add.}
}
\value{
A matrix of same size as the input, but with transformed readcounts.
}
\description{
Transforms readcount-matrix with the centered log-ratio transform.
}
\details{
This is a standard implementation of the Aitchisons centered log-ratio
transform (Aitchison 1986) for compositional data. Readcount data can be seen as compositonal data
since the total number of readcounts in a sample does not carry information, but is simply
an effect of sequencing depth. Thus, the information in the data lies in the relative 
values, not the absolute.

By transforming such data with this function, you get data who are better suited for a
number of downstream analyses, e.g. typically analyses making use of sum-of-squares type of
statistics, like PCA, PLS, ANOVA or clustering with euclidean distances.
}
\examples{

}
\references{
Aitchison J. The Statistical Analysis of Compositional Data. London, UK: Chapman & Hall; 1986.
}
\author{
Lars Snipen.
}
