% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{multinomTrain}
\alias{multinomTrain}
\title{Training multinomial model}
\usage{
multinomTrain(sequence, taxon, K = 5, col.names = FALSE, n.pseudo = 1)
}
\arguments{
\item{sequence}{Character vector of sequences.}

\item{taxon}{Character vector of taxon labels for each sequence.}

\item{K}{Word length (integer).}

\item{col.names}{Logical indicating if column names (K-mers) should be added to the trained model matrix.}

\item{n.pseudo}{Number of pseudo-counts to use (positive numerics, need not be integer). Special case -1
will only return word counts, not log-probabilities.}
}
\value{
A matrix with the multinomial probabilities, one row for each
\code{taxon} and one column for each K-mer. The sum of each row is 1.0. No
probabilities are 0 if \code{n.pseudo} > 0.0.

The matrix has an attribute \code{attr("prior",)}, that contains the relative
taxon frequencies.
}
\description{
Training the multinomial K-mer method on sequence data.
}
\details{
The training step of the multinomial method (Vinje et al, 2015) means counting K-mers
on all sequences and compute their multinomial probabilities for each taxon. 
\code{n.pseudo} pseudo-counts are added equally to all K-mers, before probabilities
are estimated. The optimal choice of \code{n.pseudo} will depend on \code{K} and the 
training data set.

Adding the actual K-mers as column names (\code{col.names = TRUE}) will slow down the
computations.

The relative taxon frequencies in the \code{taxon} input are also computed and
returned as an attribute to the probability matrix.
}
\examples{
# See examples for multinomClassify

}
\references{
Vinje, H, Liland, KH, Almøy, T, Snipen, L. (2015). Comparing K-mer based methods for
improved classification of 16S sequences. BMC Bioinformatics, 16:205.
}
\seealso{
\code{\link{KmerCount}}, \code{\link{multinomClassify}}.
}
\author{
Kristian Hovde Liland and Lars Snipen.
}
