% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclr.R
\name{rarefy}
\alias{rarefy}
\title{Rarefying readcounts}
\usage{
rarefy(readcount.mat, depth = NULL)
}
\arguments{
\item{depth}{fixed number of reads in each sample after.}

\item{readcount.mat}{matrix with readcount data.}
}
\value{
A matrix of same size as the input, but with down-sampled readcounts.
This matrix has an attribute \code{"original.depth"} with the original depths
for each sample.
}
\description{
Down-sampling readcounts to equal depths.
}
\details{
To rarefy readcounts means down-sampling the data in all samples to 
a fixed target read depth. This is sometimes done prior to some diversity analyses,
the idea being that some taxa are absent due to low depth, and this will bias
diversity estimates. After the use \code{rarefy} all samples have equal 
(low) depth.

If \code{depth = NULL} the smallest readcount among the samples is used as the
target \code{depth}, but the user may supply any other positive integer value.
Note: If \code{depth} is larger than the minimum readcount, some samples
will be up-sampled. This means they will get more reads for the taxa where they
already have reads, but still no reads for the absent taxa. 

The \code{readcount.mat} must have the samples in the rows and the taxa in the
columns. Transpose if necessary.

This function minimizes the variance in the down-sampling as follows:
For sample \code{i}, we first compute the expected readcount given target 
\code{depth}. Let
\code{Ecount = depth * read.counts[i,]/sum(read.counts[i,])}. From this we get
the base count \code{base.count = floor(Ecount)}. The remainder is 
\code{remainder = Ecount - base.count}. The remaining reads 
\code{R = depth - sum(base.count)} are finally distributed over 
the taxa using the \code{remainder} as the probabilities. Thus, only the 
latter (few) reads will vary randomly between repeated use of this function
on the same data.
}
\examples{

}
\author{
Lars Snipen.
}
