% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpfiles.R
\name{read_names_dmp}
\alias{read_names_dmp}
\alias{write_names_dmp}
\title{Read and write the names.dmp}
\usage{
read_names_dmp(filename)
write_names_dmp(names.dmp, filename)
}
\arguments{
\item{filename}{name of file to be read or written to (text).}

\item{names.dmp}{a names.dmp table (see details below).}
}
\value{
The \code{read_names_dmp} returns a \code{tibble} with the columns: \code{tax_id} (integers), \code{name_txt} (text),
\code{unique_name} (text) and \code{name_class} (text).
}
\description{
Reads and writes the file names.dmp of the NCBI Taxonomy
}
\details{
The file pair \code{names.dmp} and \code{nodes.dmp} describe a taxonomy tree. 
The \code{read_names_dmp} reads a file formatted as the \code{names.dmp} file from the NCBI Taxonomy
database (https://www.ncbi.nlm.nih.gov/taxonomy/). This is represented as a \code{tibble} in R.

The \code{write_names_dmp} will write a table with the proper columns (see below) to a file, adding 
the separators of the NCBI format.
}
\seealso{
\code{\link{read_nodes_dmp}}.
}
\author{
Lars Snipen.
}
