% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpfiles.R
\name{read_nodes_dmp}
\alias{read_nodes_dmp}
\alias{write_nodes_dmp}
\title{Read and write the nodes.dmp}
\usage{
read_nodes_dmp(filename)
write_nodes_dmp(names.dmp, filename)
}
\arguments{
\item{filename}{name of file to be read or written to.}

\item{nodes.dmp}{a nodes.dmp table (see details below).}
}
\value{
The \code{read_nodes_dmp} returns a \code{tibble} with the columns: \code{tax_id} (integers),
\code{parent_tax_id} (integers) and \code{rank} (text).
}
\description{
Reads and writes the file nodes.dmp of the NCBI Taxonomy
}
\details{
The file pair \code{names.dmp} and \code{nodes.dmp} describe a taxonomy tree.
The \code{read_nodes_dmp} reads a file formatted as the \code{nodes.dmp} file from the NCBI Taxonomy
database (https://www.ncbi.nlm.nih.gov/taxonomy/). This is represented as a \code{tibble} in R.

The \code{write_nodes_dmp} will write a table with the proper columns (see below) to a file, adding 
the separators of the NCBI format.

The \code{nodes.dmp} table downloaded from NCBI will contain many columns, but only the first 3 of them
are relevant for parsing the taxonomy tree. Only these first three columns are read and used by these functions,
additional columns are ignored.
}
\examples{

}
\seealso{
\code{\link{read_nodes_dmp}}.
}
\author{
Lars Snipen.
}
