% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{setParallel}
\alias{setParallel}
\title{Set number of parallel threads}
\usage{
setParallel(C = NULL)
}
\arguments{
\item{C}{a scalar indicating the number of threads, default = NULL (#available logical cores)}
}
\value{
NULL, returned silently.
}
\description{
Simple function to set the number of threads to use in parallel
computations. The default equals all available logical cores. An integer is
interpreted as the number of threads. A numeric < 1 is interpreted as a proportion
of the avialable logical cores.
}
\examples{
\dontrun{
setParallel() # Use all available logical cores.
}

}
