% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{subset_clade}
\alias{subset_clade}
\title{Subset the taxonomy tree from the root}
\usage{
subset_clade(root.tax.id, nodes.dmp)
}
\arguments{
\item{root.tax.id}{the tax_id of the clade root (integer).}

\item{nodes.dmp}{a nodes.dmp table.}
}
\value{
A subset of the table \code{nodes.dmp} containing the clade that descends from \code{root.tax.id}.
}
\description{
Retrieves a clade from the taxonomy tree.
}
\details{
This function retrieves a clade from the taxonomy tree listed in the
table \code{nodes.dmp} (see \code{\link{read_nodes_dmp}}). The clade starts at the single taxon specified in
\code{root_tax_id} and contains all descending branches.

If you want to subset the taxonomy tree based on leaf nodes, see \code{\link{subset_tree}}.
}
\seealso{
\code{\link{read_nodes_dmp}}, \code{\link{subset_tree}}.
}
\author{
Lars Snipen.
}
