% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitree.R
\name{subset_tree}
\alias{subset_tree}
\title{Subset the taxonomy tree from leaves}
\usage{
subset_tree(leaf_tax_id, nodes.dmp)
}
\arguments{
\item{leaf.tax.id}{a vector of tax_id (integers) of the branch leaf/leaves.}

\item{nodes.dmp}{a nodes.dmp table.}
}
\value{
A subset of the table \code{nodes.dmp} containing the tax_id's of the branches
ending at \code{leaf_tax_id}.
}
\description{
Retrieves a sub-tree from the taxonomy tree.
}
\details{
This function retrieves a sub-tree by starting at the vector of specified
\code{leaf_tax_id} and collect all branches ending at these nodes. Thus, it will in general
not be a clade. If you want to subset an entire clade, see \code{\link{subset_clade}}.
}
\examples{

}
\seealso{
\code{\link{read_nodes_dmp}}, \code{\link{subset_clade}}.
}
\author{
Lars Snipen.
}
