% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{small.16S}
\alias{small.16S}
\title{A small example data set}
\usage{
data(small.16S)
}
\description{
A \code{tibble} object (\code{data.frame}) with some 16S sequences with taxon information.
}
\details{
This is a \code{tibble} object (\code{data.frame}) with 71 sequences used in some examples.
The taxonomy information for each sequence follows the ConTax format, see the 
\code{\link{microcontax}} package for more details.
}
\examples{
data(small.16S)
str(small.16S)

}
\author{
Hilde Vinje, Kristian Hovde Liland, Lars Snipen.
}
